Refactoring: Extract creation of index template requests into own factory
